requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.25.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
